fix(gui): wrap Demo Noise button strips instead of clipping at rail width (#4518)
The Demo Noise applet packed six labeled buttons per strip into flat
QHBoxLayouts. Docked, the applet rail's widget-resizable scroll area
(horizontal scrolling off) hands the layout less than its minimum, so
the buttons compress past their hint and the centred labels clip at
both ends to unreadable mid-word fragments.

Same failure mode SpotHub's band filters hit in #4157, and the cure is
the same FlowLayout that fix introduced — promoted here from file-local
in DxClusterDialog.cpp to a shared src/gui/FlowLayout.h (moved verbatim)
and used for both DemoApplet strips. Buttons now keep their natural
width and wrap to additional rows as needed, at any width.
